apps/api.backup/requirements.txt
fastapi
uvicorn
motor
beanie
tensorflow
numpy
pillow
python-multipart
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pytest
httpx
pandas
scipy
email-validator
python-dotenv
google-generativeai
